chore(refactor): extract OAuth + /mcp into routes/oauth-mcp.ts
Sixth bucket of the index.ts route extraction.

server/src/routes/oauth-mcp.ts (new) — seven endpoints:
  GET  /.well-known/oauth-protected-resource/mcp
  GET  /.well-known/oauth-authorization-server
  POST /oauth/register
  GET  /oauth/authorize
  POST /oauth/token
  *    /mcp, /mcp/, /mcp?…       (StreamableHTTPServerTransport)
  GET  /api/mcp/status

The OAuth dance is intentionally a no-op — Oyster runs localhost-only and
doesn't authenticate clients; the discovery + register + authorize + token
endpoints only exist so MCP SDK clients that probe for them don't fail.

Drops the duplicate `json` helper that lived inside handleHttpRequest
(audit comment #15) — the route module uses ctx.sendJson instead.

Imports cleanup: createMcpServer, recordExternalRequest +
listExternalClients + externalClientCount + lastConnectedAt, and
StreamableHTTPServerTransport are no longer referenced from index.ts;
moved into routes/oauth-mcp.ts where they're now used.

Cumulative: index.ts is now 1292 lines — down 926 lines (-42%) from the
pre-audit baseline of 2218.

No behavioural change.
